Understand contravariance
In the previous commit I defined the IReadRepository interface with a Covariant type parameter, IReadRepository<out T>, this means that the type can be less specific on the interface.

But now with the Manager I need a more specific type, a Contravariant type, to create a Contravariant type parameter you write the in keyword before the type parameter,<in T>, this works if the type parameter is not used as a return value in the generic interfaces.